requirements.txt
pymupdf
pdfplumber
regex
opencv-python
pytesseract
Pillow
streamlit
numpy
matplotlib
json5